+++ src/routes/routes.constant.ts
@@ -22,6 +22,7 @@ export const NAVIGATION_ROUTES = {
 
   DOCTOR_LIST: "/doctor-list",
   DOCTOR_PROFILE: "/doctor-profile",
+  DOCTOR_PROFILES: "/doctor-profiles",
   APPOINTMENTS: "appointments",
   FOLLOWUP: "followup",
   PATIENT_HISTORY: "patient-history",